kernel: Decouple from full init/common.cpp
...by splitting the libbitcoinkernel-used portion of init/common.cpp off
   to kernel/init/common.cpp

Notes for reviewers:
    - I strongly recommend reviewing with the following git-diff flags:
          --patience --color-moved=dimmed-zebra
    - See the diff for src/Makefile.am to see the overall effect.
