+++ src/init/common.cpp
@@ -8,36 +8,15 @@
 
 #include <clientversion.h>
 #include <compat/sanity.h>
-#include <crypto/sha256.h>
 #include <key.h>
 #include <logging.h>
 #include <node/ui_interface.h>
-#include <pubkey.h>
 #include <random.h>
 #include <util/system.h>
 #include <util/time.h>
 #include <util/translation.h>
 
-#include <memory>
-
-static std::unique_ptr<ECCVerifyHandle> globalVerifyHandle;
-
 namespace init {
-void SetGlobals()
-{
-    std::string sha256_algo = SHA256AutoDetect();
-    LogPrintf("Using the '%s' SHA256 implementation\n", sha256_algo);
-    RandomInit();
-    ECC_Start();
-    globalVerifyHandle.reset(new ECCVerifyHandle());
-}
-
-void UnsetGlobals()
-{
-    globalVerifyHandle.reset();
-    ECC_Stop();
-}
-
 bool SanityChecks()
 {
     if (!ECC_InitSanityCheck()) {
